My graph dynamique ok
