+++ SofIA/extension/components/pages/graph-tabs/ResonanceTab.tsx
@@ -1,0 +1,14 @@
+const ResonanceTab = () => {
+  return (
+    <div className="triples-container">
+      <div className="empty-state">
+        <p>No resonance data available</p>
+        <p className="empty-subtext">
+          Resonance patterns will appear here when detected
+        </p>
+      </div>
+    </div>
+  )
+}
+
+export default ResonanceTab
