+++ SofIA/extension/components/pages/graph-tabs/SignalsTab.tsx
@@ -1,0 +1,14 @@
+const SignalsTab = () => {
+  return (
+    <div className="triples-container">
+      <div className="empty-state">
+        <p>No triples registered yet</p>
+        <p className="empty-subtext">
+          Your validated triplets will appear here once stored on-chain
+        </p>
+      </div>
+    </div>
+  )
+}
+
+export default SignalsTab
