+++ SofIA/extension/components/styles/MyGraphPage.css
@@ -2,6 +2,9 @@
 
 .page {
   text-align: center;
+  overflow-x: hidden;
+  max-width: 100vw;
+  box-sizing: border-box;
 }
 
 .section-title {
@@ -30,6 +33,7 @@
   cursor: pointer;
   font-size: 16px;
   font-weight: 400;
+  font-family: 'Gotu', cursive;
   transition: all 0.3s ease;
   position: relative;
 }
@@ -61,7 +65,13 @@
 .triples-container {
   display: flex;
   flex-direction: column;
-  gap: 20px;
+  gap: 10px;
+  max-width: min(600px, 100%);
+  margin: 0 auto;
+  text-align: left;
+  padding: 0 16px;
+  box-sizing: border-box;
+  overflow-x: hidden;
 }
 
 .subsection-title {
@@ -70,6 +80,20 @@
   font-weight: 500;
   color: #F2DED6;
   margin-bottom: 10px;
+}
+
+.loading-state {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 40px;
+  font-family: 'Montserrat', sans-serif;
+  font-size: 16px;
+  color: #C7866C;
+  background-color: rgba(255, 255, 255, 0.05);
+  border-radius: 12px;
+  backdrop-filter: blur(5px) saturate(100%);
+  -webkit-backdrop-filter: blur(5px) saturate(100%);
 }
 
 /* Utilise .empty-state de Global.css avec override spécifique */
@@ -130,24 +154,16 @@
   margin: 8px 0;
 }
 
-.triples-container {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  max-width: 600px;
-  margin: 0 auto;
-  text-align: left;
-}
 
 .echo-card {
   background-color: rgba(0, 0, 0, 0.8);
   border-radius: 12px;
   padding: 20px;
   margin-bottom: 20px;
-  backdrop-filter: blur(30px) saturate(100%);
-  -webkit-backdrop-filter: blur(30px) saturate(100%);
+  backdrop-filter: blur(50px) saturate(100%);
+  -webkit-backdrop-filter: blur(50px) saturate(100%);
   border: 1px solid rgba(255, 255, 255, 0.1);
-  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.2), inset 0 1px 0 rgba(255, 255, 255, 0.1);
+  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.5), inset 0 1px 0 rgba(255, 255, 255, 0.5);
   transition: all 0.3s ease;
 }
 
@@ -205,7 +221,7 @@
 .add-button:hover {
   background-color: rgba(199, 134, 108, 0.4);
   border-color: rgba(199, 134, 108, 0.5);
-  transform: scale(1.05);
+  transform: scale(1.02);
 }
 
 .quick-action-button {
@@ -221,11 +237,85 @@
 
 .quick-action-button:hover,
 .quick-action-button.hovered {
-  transform: scale(1.05);
+  transform: scale(1.02);
 }
 
 .quick-action-icon {
-  width: 54px;
-  height: 18px;
+  width: 72px;
+  height: 24px;
   display: block;
+}
+
+/* Triplet Details Section - SofIA Design System */
+.triplet-details {
+  margin-top: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  backdrop-filter: blur(10px) saturate(100%);
+  -webkit-backdrop-filter: blur(10px) saturate(100%);
+  transition: all 0.3s ease;
+}
+
+.triplet-detail-section {
+  padding: 16px;
+  background-color: rgba(251, 247, 245, 0.05); /* 50 avec transparence */
+  border-radius: 8px;
+  border: 1px solid rgba(242, 222, 214, 0.1); /* 200 avec transparence */
+  transition: all 0.3s ease;
+}
+
+.triplet-detail-section:hover {
+  background-color: rgba(251, 247, 245, 0.08);
+  border-color: rgba(199, 134, 108, 0.3); /* 500 avec transparence */
+  transform: translateY(-1px);
+}
+
+.triplet-detail-title {
+  font-family: 'Gotu', cursive; /* Titres selon design system */
+  font-size: 16px;
+  font-weight: 600;
+  color: #C7866C; /* 500 */
+  margin: 0 0 12px 0;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.triplet-detail-name {
+  font-family: 'Montserrat', sans-serif; /* Texte selon design system */
+  font-size: 14px;
+  font-weight: 600;
+  color: #FBF7F5; /* 50 */
+  margin: 0 0 8px 0;
+  line-height: 1.4;
+}
+
+.triplet-detail-description {
+  font-family: 'Montserrat', sans-serif; /* Texte selon design system */
+  font-size: 13px;
+  font-weight: 400;
+  color: #F2DED6; /* 200 */
+  margin: 0 0 8px 0;
+  line-height: 1.5;
+  opacity: 0.9;
+}
+
+.triplet-detail-url {
+  font-family: 'Montserrat', sans-serif; /* Texte selon design system */
+  font-size: 12px;
+  color: #F2DED6; /* 200 */
+  text-decoration: underline;
+  transition: all 0.3s ease;
+  word-break: break-all;
+  overflow-wrap: break-word;
+  display: inline-block;
+  margin-top: 4px;
+  max-width: 100%;
+}
+
+.triplet-detail-url:hover {
+  color: #C7866C; /* 500 */
+  text-decoration: none;
+  transform: translateX(2px);
 }
